move get_index to transform
