+++ rasterio/windows.py
@@ -12,6 +12,8 @@ import math
 from affine import Affine
 import numpy as np
 
+from rasterio.transform import get_index
+
 
 def iter_args(function):
     """Decorator to allow function to take either *args or
@@ -241,42 +243,6 @@ def evaluate(window, height, width):
     return (r_start, r_stop), (c_start, c_stop)
 
 
-def get_index(x, y, transform, op=math.floor, precision=6):
-    """
-    Returns the (row, col) index of the pixel containing (x, y) given a
-    coordinate reference system.
-
-    Parameters
-    ----------
-    x : float
-        x value in coordinate reference system
-    y : float
-        y value in coordinate reference system
-    transform : Affine
-        Coefficients mapping pixel coordinates to coordinate reference system.
-    op : function
-        Function to convert fractional pixels to whole numbers (floor, ceiling,
-        round)
-    precision : int
-        Decimal places of precision in indexing, as in `round()`.
-
-    Returns
-    -------
-    row : int
-        row index
-    col : int
-        col index
-    """
-    # Use an epsilon, magnitude determined by the precision parameter
-    # and sign determined by the op function: positive for floor, negative
-    # for ceil.
-    eps = 10.0**-precision * (1.0 - 2.0*op(0.1))
-    fcol, frow = ~transform * (x + eps, y - eps)
-    col = int(op(fcol))
-    row = int(op(frow))
-    return row, col
-
-
 def shape(window, height=-1, width=-1):
     """Returns shape of a window.
 
@@ -289,5 +255,5 @@ def shape(window, height=-1, width=-1):
 
 def window_index(window):
     # "window_" is necessary here to redundancy to disambiguate
-    # from get_index which provides src.index
+    # from transform.get_index and src.index
     return tuple(slice(*w) for w in window)
